# contracts/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      description: Stable error envelope returned by every ULOO Core endpoint.
      properties:
        code:
          description: Stable machine-readable error code
          title: Code
          type: string
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
        message:
          description: Safe human-readable error message
          title: Message
          type: string
        request_id:
          title: Request Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - code
      - message
      - request_id
      - trace_id
      title: ErrorResponse
      type: object
...
    RunCreate:
      properties:
        input:
          anyOf:
          - maxLength: 100000
            minLength: 1
            type: string
          - type: 'null'
          title: Input
      title: RunCreate
      type: object
    RunDetailResponse:
      properties:
        agno_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agno Run Id
        configuration_snapshot:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Configuration Snapshot
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        events:
          items:
            $ref: '#/components/schemas/RunEventResponse'
          title: Events
          type: array
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        id:
          title: Id
          type: string
        input_text:
          title: Input Text
          type: string
        is_mock:
          title: Is Mock
          type: boolean
        output:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output
        plan_id:
          title: Plan Id
          type: string
        runtime_type:
          title: Runtime Type
          type: string
        started_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Started At
        status:
          title: Status
          type: string
        task_id:
          title: Task Id
          type: string
        team_id:
          title: Team Id
          type: string
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
        usage:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Usage
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - id
      - workspace_id
      - task_id
      - plan_id
      - team_id
      - status
      - runtime_type
      - is_mock
      - input_text
      title: RunDetailResponse
      type: object
    RunEventResponse:
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        event_type:
          title: Event Type
          type: string
        id:
          title: Id
          type: string
        payload:
          additionalProperties: true
          title: Payload
          type: object
        run_id:
          title: Run Id
          type: string
        sequence:
          minimum: 1.0
          title: Sequence
          type: integer
        source_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Id
        source_type:
          title: Source Type
          type: string
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      required:
      - id
      - run_id
      - sequence
      - event_type
      - source_type
      title: RunEventResponse
      type: object
    RunListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RunResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - offset
      - limit
      title: RunListResponse
      type: object
...
info:
  title: ULOO Core
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/runs:
    get:
      description: List persisted runs in the current workspace.
      operationId: list_runs_api_v1_runs_get
      parameters:
      - in: query
        name: task_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Task Id
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: List Runs
      tags:
      - runs
  /api/v1/runs/{run_id}:
    get:
      description: Return one run and its ordered observable events.
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          format: uuid
          title: Run Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Run
      tags:
      - runs
  /api/v1/runs/{run_id}/events:
    get:
      description: Return ordered events after a sequence cursor; SSE is added in the next stage.
      operationId: get_run_events_api_v1_runs__run_id__events_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          format: uuid
          title: Run Id
          type: string
      - in: query
        name: after
        required: false
        schema:
          default: 0
          minimum: 0
          title: After
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RunEventResponse'
                title: Response Get Run Events Api V1 Runs  Run Id  Events Get
                type: array
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Run Events
      tags:
      - runs
...
  /api/v1/tasks/{task_id}/approve:
    post:
      description: Approve the task's current pending plan before execution.
      operationId: approve_task_plan_api_v1_tasks__task_id__approve_post
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          format: uuid
          title: Task Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  anyOf:
                  - type: string
                  - type: integer
                title: Response Approve Task Plan Api V1 Tasks  Task Id  Approve Post
                type: object
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Approve Task Plan
      tags:
      - runs
...
  /api/v1/tasks/{task_id}/runs:
    post:
      description: Execute the approved plan through a real Agno Team and persist safe events.
      operationId: start_task_run_api_v1_tasks__task_id__runs_post
      parameters:
      - in: path
        name: task_id
        required: true
        schema:
          format: uuid
          title: Task Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RunCreate'
              - type: 'null'
              title: Body
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Start Task Run
      tags:
      - runs
...